Refactor searchAuthor.js to generic search.js
The searchAuthor.js was renamed to search.js to specify more general search operations. The search command was also updated for more flexibility, enabling users to provide any author's name to conduct a search instead of a pre-defined one. Additionally, the number of reported search results was increased from 7 to 10.