Update /author and /search commands
The footer and additional fields are added to the /author command interface, improving its features and warning users that it is still under development. The /search command also updated by adding a development warning in the 'footer'. The help.js file was also updated to reflect the development status of the /author command.